IndexedDB/tests/submissions/Microsoft/idb_webworkers.htm

<!DOCTYPE html>
<html>
<head>
    <title>IndexedDB inside of a WebWorker </title>
    <script type="text/javascript" src="support.js"></script>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script type="text/javascript">
        var t = async_test(); 

        function RunTest() {
            var worker = new Worker("idbworker.js");
            worker.onmessage = t.step_func( function (event) {
                assert_equals(event.data, true, 'self.indexedDB');
                t.done();
            });
            worker.postMessage(1);
        }
        
        t.step( RunTest );

    </script>
</head>
<body>
    <div id="log"></div>
</body>
</html>
